Moved GVG Single Player Task back
Did not mean to move this into the player package earlier
